irqchip: Add driver for Axxia GIC
Each of the CPU clusters in the AXM55xx device has an ARM GIC. The masking and
unmasking of interrupts via the GIC distributor interface must be done from
within the relevant cluster.

To handle IPIs between clusters, four any-CPU-to-any-CPU peripheral interrupts
are used. The kernel IPIs are multiplexed on (currently one) of these
interrupts.

Signed-off-by: Anders Berg <anders.berg@lsi.com>