ARM: dts: axxia: Add property expected by boot loader
Add a 'frequency' property to be able to use the existing boot loader. This
property is expected by boot loader to be present in the fdt. The result if
it's not found, is a fatal error. This fix in the kernel DTS is to be
considered temporary until the boot loader is updated to behave properly.

Signed-off-by: Anders Berg <anders.berg@lsi.com>